cmd/dolt: address review feedback on error sentinel and OS build files
